feat: Initial packit setup
This commit contains the specfile for building the official package for Fedora
with a Packit setup.

Ultimately, a unified specfile is targeted for Fedora and any other rpm-based
distributions, e.g. openEuler.

And Packit(https://packit.dev/) is a tool for maintaining specfile within
upstream source. It requires a simple config file(.packit.yaml).

Log:
Signed-off-by: Robin Lee <cheeselee@fedoraproject.org>
Change-Id: Ibf9275a06f26678e993f5942cf28e38867370322
Reviewed-on: http://gerrit.uniontech.com/c/dde-launcher/+/8542
Reviewed-by: zhangjide <zhangjide@uniontech.com>
Reviewed-by: zhangqiuliang <zhangqiuliang@uniontech.com>
Tested-by: zhangjide <zhangjide@uniontech.com>