fix: incorrect trash status when deleting file in other partition
The old implementation only monitor the trash folder in user's
partition, thus the trash folder in other partition will not trached. It
will cause the launcher icon status out-of-sync with the real trash
status.

This patch switch to use gio to monitor the trash status (might need to
port to dtkio when it get a stable release).

Issue: https://github.com/linuxdeepin/developer-center/issues/4478
Log: Fix the trash icon status might out-of-sync in some cases